updated examples
