+++ examples/deps.ts
@@ -1,8 +1,12 @@
-export * from "https://deno.land/x/dxcfg@v0.0.6/mod.ts";
-export * from "https://deno.land/x/dxcfg@v0.0.6/kubernetes/mod.ts";
-export * from "https://deno.land/x/dxcfg@v0.0.6/kubernetes/gen/mod.ts";
+export * from "https://deno.land/x/dxcfg@v0.0.7/mod.ts";
+export * from "https://deno.land/x/dxcfg@v0.0.7/kubernetes/mod.ts";
+export * from "https://deno.land/x/dxcfg@v0.0.7/kubernetes/gen/mod.ts";
 export {
   assertEquals,
   assertObjectMatch,
   assertThrows,
 } from "https://deno.land/std@0.114.0/testing/asserts.ts";
+
+import HandlebarsJS from "https://dev.jspm.io/handlebars@4.7.6";
+export { HandlebarsJS };
+export { dir } from "../fs.ts";
