+++ examples/kubernetes/chart-simple/chart.yaml
@@ -1,25 +1,22 @@
-- file: helloworld-dep-deployment.yaml
-  value:
-    apiVersion: apps/v1
-    kind: Deployment
-    metadata:
-      name: helloworld-dep
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: demo-dep
+spec:
+  template:
+    labels:
+      app: hello
     spec:
-      template:
-        labels:
-          app: hello
-        spec:
-          containers:
-            hello:
-              image: 'weaveworks/helloworld:v1'
-- file: helloworld-svc-service.yaml
-  value:
-    apiVersion: v1
-    kind: Service
-    metadata:
-      name: helloworld-svc
-      labels:
-        app: hello
-    spec:
-      selector:
-        app: hello
+      containers:
+        hello:
+          image: 'weaveworks/helloworld:v2'
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: demo-svc
+  labels:
+    app: hello
+spec:
+  selector:
+    app: hello
