--- examples/kubernetes/chart-simple/test.cmd
@@ -1,2 +1,0 @@
-npm ci
-jk generate --stdout ./index.js
