--- examples/kubernetes/chart-simple/test.expected
@@ -1,22 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: helloworld-dep
-spec:
-  template:
-    labels:
-      app: hello
-    spec:
-      containers:
-        hello:
-          image: weaveworks/helloworld:v1
----
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    app: hello
-  name: helloworld-svc
-spec:
-  selector:
-    app: hello
