--- examples/kubernetes/chart-simple/test.param.cmd
@@ -1,2 +1,0 @@
-npm ci
-jk generate --stdout ./index.js -p values.name=demo -p values.image.tag=v2
