+++ examples/kubernetes/chart-templated/chart.yaml
@@ -1,0 +1,22 @@
+|-
+  apiVersion: apps/v1
+  kind: Deployment
+  metadata:
+    name: 'helloworld-dep'
+  spec:
+    template:
+      labels:
+        app: 'hello'
+        spec:
+          containers:
+          - name: hello
+            image: 'weaveworks/helloworld:v1'
+---
+|-
+  apiVersion: v1
+  kind: Service
+  metadata:
+    name: 'helloworld-svc'
+  spec:
+    selector:
+      app: 'hello'
