+++ examples/kubernetes/chart-templated/defaults.yaml
