+++ examples/kubernetes/chart-templated/index.js
@@ -1,6 +1,15 @@
-// Example of a Helm chart analogue, using handlebars
+import { param, read, write, Format, dir } from '../../deps.ts';
+import { HandlebarsJS } from "../../deps.ts";
+const compile = (HandlebarsJS).compile;
 
-import * as param from '@jkcfg/std/param';
-import chart from './chart';
+const resources = (values) => {
+    return dir("./templates").files.map(f => {
+        const template = compile(Deno.readTextFileSync(f.path))
+        return template({ values })
+    })
+}
 
-export default chart(param);
+const values = await param.all();
+const res = resources(values)
+await write(resources(values), 'chart.yaml', { format: Format.MULTI_YAML })
+console.log(await read('chart.yaml'))
