--- examples/kubernetes/chart-templated/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@jkcfg/kubernetes-examples-chart-templated",
-  "version": "1.0.0",
-  "description": "Example of using jk to generate Kubernetes config",
-  "main": "index.js",
-  "dependencies": {
-    "@jkcfg/kubernetes": "^0.5.0",
-    "handlebars": "^4.5.3"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "generate": "jk generate --stdout ./index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jk/examples.git"
-  },
-  "keywords": [
-    "kubernetes",
-    "jk",
-    "configuration"
-  ],
-  "author": "The jk authors",
-  "license": "Apache License 2.0",
-  "bugs": {
-    "url": "https://github.com/jk/examples/issues"
-  },
-  "homepage": "https://github.com/jk/examples/kubernetes#readme"
-}
