+++ examples/kubernetes/chart-templated/templates/deployment.yaml
