+++ examples/kubernetes/chart-templated/templates/service.yaml
